refactor (test request)
change all loops to recursive functions; add async-await; add URL object; edit naming; remove modules.js and replace it with requests.js and extra.js; style: remove unnecessary semicolons
